divided contacts-controller
