+++ controllers/contacts/getAllContacts.js
@@ -1,0 +1,8 @@
+const { Contact } = require("../../models/contact");
+
+const getAllContacts = async (req, res) => {
+    const result = await Contact.find({}, "-createdAt -updatedAt");
+    res.json(result)
+}
+
+module.exports = getAllContacts;
